pyproject.toml
[tool.poetry]
name = "djangofw"
version = "0.1.0"
description = "kw7.drf"
authors = ["max"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
django = "4.2"
psycopg2 = "^2.9.9"
djangorestframework = "^3.15.2"
ipython = "^8.25.0"
pillow = "^10.3.0"
black = "^24.4.2"
djangorestframework-simplejwt = "^5.3.1"
celery = "^5.4.0"
redis = "^5.0.6"
django-celery-beat = "^2.6.0"
requests = "^2.32.3"
coverage = "^7.5.4"
django-filter = "^24.2"
drf-yasg = "^1.21.7"
setuptools = "^70.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
